Revert "Revert "Revert "Fix: RAITA-200 Use another lambda to invoke the zipping process (#119)" (#120)" (#121)"
This reverts commit 44d8af732237a284af1f73ba141121e347fc448e.
